moving Reader to a single bundle
remove wildcards

Move all page and renders into parent reader.

es6ify controller/index.js

switch to renderWithReduxStore for everything

hoisting all of the imports

also removed an unneeded feature flag

fixed primary/secondary garbage for fullpost/main stuff

remove preloader for Reader
